fix(routes): resolve 404/403/500 errors on analytics, insights, audit, permissions
Routes & Templates:
- Change /finance/insights to /insights route
- Change /finance/audit/timeline to /audit route
- Create permissions/list.html template
- Update form actions and export links in templates

Analytics/Insights SQL:
- Add explicit ::bigint casts to nullable branch_id parameters
- Fix 'could not determine data type of parameter' PostgreSQL errors
- Update both source SQL and generated Go files

Permissions:
- Add migration 000015 for finance analytics permissions
- Add finance.view_insights, finance.view_audit, permissions.view to seed
- Assign new permissions to admin role

Sidebar:
- Fix double-active state on /analytics/kpi
- Use exact match only for nav-item active class

Workflow:
- Add fix-backend.md workflow for backend debugging

+++ internal/insights/db/insights.sql.go
@@ -16,7 +16,7 @@ SELECT period,
 FROM mv_pl_monthly
 WHERE period BETWEEN $1 AND $2
   AND company_id = $3
-  AND ($4 IS NULL OR branch_id = $4)
+  AND ($4::bigint IS NULL OR branch_id = $4::bigint)
 GROUP BY period
 ORDER BY period
 `
